FIX broke label_probagation example, now fixed it again.

+++ sklearn/datasets/samples_generator.py
@@ -459,7 +459,7 @@ def make_circles(n_samples=100, shuffle=True, noise=None, random_state=None,
     if factor > 1 or factor < 0:
         raise ValueError("'factor' has to be between 0 and 1.")
 
-    n_samples_out = n_samples / (1 + factor)
+    n_samples_out = int(n_samples / float(1 + factor))
     n_samples_in = n_samples - n_samples_out
 
     generator = check_random_state(random_state)
